fixed "double" seconds in the PZP plugin

+++ src/powerzonepack.cpp
@@ -119,9 +119,9 @@ void powerzonepack::search_workout_onfinish(QNetworkReply* reply)
     for(int i=1; i<power_graph.count(); i++)
     {
         trainrow r;
-        int sec = power_graph.at(i).toObject()["seconds"].toInt();
+        double sec = power_graph.at(i).toObject()["seconds"].toDouble();
         QTime seconds(0,0,0,0);
-        seconds = seconds.addSecs(sec);
+        seconds = seconds.addSecs((int)sec);
         r.duration = QTime(0,0,lastSeconds.msecsTo(seconds) / 1000,0);
         r.power = power_graph.at(i - 1).toObject()["power_ratio"].toDouble() * settings.value("ftp", 200.0).toDouble();
         lastSeconds = seconds;
